+++ packages/platforms/react-native/lib/client.ts
@@ -1,60 +1,48 @@
-import { createClient } from '@bugsnag/core-performance'
-import createFetchDeliveryFactory from '@bugsnag/delivery-fetch-performance'
-import { createXmlHttpRequestTracker } from '@bugsnag/request-tracker-performance'
-import { AppRegistry, AppState } from 'react-native'
-import { FileSystem } from './persistence/file-native'
-import { AppStartPlugin, NetworkRequestPlugin } from './auto-instrumentation'
+/* eslint-disable @typescript-eslint/consistent-type-assertions */
 import createClock from './clock'
-import createSchema from './config'
-import createIdGenerator from './id-generator'
-import NativeBugsnagPerformance from './native'
-import persistenceFactory from './persistence'
-import { platformExtensions } from './platform-extensions'
-import resourceAttributesSourceFactory from './resource-attributes-source'
-import createRetryQueueFactory from './retry-queue'
-import { createSpanAttributesSource } from './span-attributes-source'
-import createBrowserBackgroundingListener from './backgrounding-listener'
-import { ReactNativeSpanFactory } from './span-factory'
+import type { ReactNativeConfiguration } from './config'
+import { createReactNativeClient } from './create-client'
+import type { PlatformExtensions } from './platform-extensions'
+import type { BugsnagPerformance } from '@bugsnag/core-performance'
 
-// this is how some internal react native code detects whether the app is running
-// in the remote debugger:
-// https://github.com/facebook/react-native/blob/e320ab47cf855f2e5de74ea448ec292cf0bbb29a/packages/react-native/Libraries/Utilities/DebugEnvironment.js#L15
-// there's no public api for this so we use the same approach
+// Singleton client instance shared across the entire application.
+// Will be lazily created on first property access to the BugsnagPerformance proxy below,
+// OR can be pre-configured by calling registerClient()
+let clientInstance: unknown
 
-// @ts-expect-error Element implicitly has an 'any' type because type 'typeof globalThis' has no index signature.
-const isDebuggingRemotely = !global.nativeCallSyncHook && !global.RN$Bridgeless
+// Sets the singleton BugsnagPerformance client instance.
+// Upstream libraries such as Expo MUST call this to pre-configure the client instance upfront before any usage.
+export function registerClient<C extends ReactNativeConfiguration = ReactNativeConfiguration, T = PlatformExtensions> (client: BugsnagPerformance<C, T>): void {
+  clientInstance = client
+}
 
 const clock = createClock(performance)
-const isDevelopment = __DEV__
 const appStartTime = clock.now()
-const deliveryFactory = createFetchDeliveryFactory(fetch, clock)
-const spanAttributesSource = createSpanAttributesSource()
-const deviceInfo = !isDebuggingRemotely ? NativeBugsnagPerformance.getDeviceInfo() : undefined
-const persistence = persistenceFactory(FileSystem, deviceInfo)
-const resourceAttributesSource = resourceAttributesSourceFactory(persistence, deviceInfo)
-const backgroundingListener = createBrowserBackgroundingListener(AppState)
 
-// React Native's fetch polyfill uses xhr under the hood, so we only track xhr requests
-const xhrRequestTracker = createXmlHttpRequestTracker(XMLHttpRequest, clock)
+// Proxy wrapper for lazy initialization of the BugsnagPerformance client.
+// On first property access:
+// - Uses pre-configured client from registerClient() if available, otherwise creates a new client
+// - Copies all client properties to the proxy target and removes the get trap
+// - All subsequent accesses go directly to the target, bypassing the proxy for optimal performance
+const client = new Proxy({} as BugsnagPerformance<ReactNativeConfiguration, PlatformExtensions>, {
+  get (target, prop) {
+    // React uses the $$typeof property to identify React elements - we return undefined
+    // to avoid triggering client creation during React's internal type checking
+    if (prop === '$$typeof') return undefined
 
-const idGenerator = createIdGenerator(isDebuggingRemotely)
-const BugsnagPerformance = createClient({
-  isDevelopment,
-  backgroundingListener,
-  clock,
-  deliveryFactory,
-  idGenerator,
-  persistence,
-  plugins: (spanFactory, spanContextStorage) => [
-    new AppStartPlugin(appStartTime, spanFactory as ReactNativeSpanFactory, clock, AppRegistry),
-    new NetworkRequestPlugin(spanFactory, spanContextStorage, xhrRequestTracker)
-  ],
-  resourceAttributesSource,
-  schema: createSchema(isDevelopment),
-  spanAttributesSource,
-  retryQueueFactory: createRetryQueueFactory(FileSystem),
-  spanFactory: ReactNativeSpanFactory,
-  platformExtensions: (spanFactory, spanContextStorage) => platformExtensions(appStartTime, clock, spanFactory as ReactNativeSpanFactory, spanContextStorage)
+    // initialize the client instance here if it hasn't already been set via registerClient()
+    if (!clientInstance) {
+      clientInstance = createReactNativeClient({ appStartTime, clock })
+    }
+
+    // Now that the client has been initialized, we can add all its properties to the proxy target
+    // and remove the get trap in order to bypass the proxy for future accesses
+    const client = clientInstance as BugsnagPerformance<ReactNativeConfiguration, PlatformExtensions>
+    Object.defineProperties(target, Object.getOwnPropertyDescriptors(client))
+    delete this.get
+
+    return client[prop as keyof typeof client]
+  }
 })
 
-export default BugsnagPerformance
+export default client