+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- (react-native) Introduced a lazy-loading singleton client to ensure a single client instance per app [#748](https://github.com/bugsnag/bugsnag-js-performance/pull/748)
+
+## [v3.2.0] (2025-11-13)
+
+### Added
+
+- (react-native) Added `createReactNativeClient` factory method for internal use by upstream libraries [#730](https://github.com/bugsnag/bugsnag-js-performance/pull/730)
+
+### Changed
+
+- Amend secondary endpoint to `bugsnag.smartbear.com` [#737](https://github.com/bugsnag/bugsnag-js/pull/737)
+
 ## [v3.1.0] (2025-09-24)
 
 ### Added